UTILS: Add uint16_t casts
Several sizes are guaranteed to be 16 bits (TPM2B) when scripting, but
libraries used size_t, which can be larger.  (An application accepting
untrusted input would verify the size before the cast.

+++ utils/certifycreation.c
@@ -352,7 +352,7 @@ int main(int argc, char *argv[])
 	}
     }
     if (rc == 0) {
-	in.creationHash.t.size = length;
+	in.creationHash.t.size = (uint16_t)length;
 	memcpy(in.creationHash.t.buffer, buffer, length);
     }
     /* Start a TSS context */
